Split out logger activation code; logger.ts now contains just the Logger interface and instance
